Refactoring the unit test + increasing coverage + decreasing run time
