+++ api/app/tests/apps/api/v1/json_schema/get_signals_v1_public_terms_categories.json
