+++ api/app/tests/apps/api/v1/json_schema/post_signals_v1_public_signals_attachment.json
