+++ api/app/tests/apps/api/v1/request_data/create_initial.json
@@ -22,7 +22,7 @@
         "extra_properties": {}
     },
     "category": {
-        "sub_category": "http://example.com/placeholder/replace/in/tests"
+        "sub_category": "/signals/v1/public/terms/categories/overig/sub_categories/overig"
     },
     "reporter": {
         "email": "melder@example.com",
