+++ api/app/tests/apps/api/v1/request_data/update_location.json
@@ -15,7 +15,7 @@
             "postcode": "1012AA",
             "woonplaats": "Amsterdam"
         },
-        "extra_properties": {},
+        "extra_properties": null,
         "stadsdeel": "A",
         "buurt_code": "A01a"
     }
