+++ api/app/tests/apps/api/v1/test_pdf.py
@@ -1,5 +1,3 @@
-from django.urls import reverse
-
 from tests.apps.signals.factories import SignalFactory
 from tests.test import SIAReadWriteUserMixin, SignalsBaseApiTestCase
 
@@ -10,10 +8,8 @@ class TestPDFView(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
 
     def test_get_pdf(self):
         self.client.force_authenticate(user=self.sia_read_write_user)
-        response = self.client.get(path=reverse(
-            'v1:signal-pdf-download',
-            kwargs={'pk': self.signal.id})
-        )
+        url = '/signals/v1/private/signals/{}/pdf'.format(self.signal.pk)
+        response = self.client.get(path=url)
 
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.get('Content-Type'), 'application/pdf')
@@ -24,19 +20,14 @@ class TestPDFView(SIAReadWriteUserMixin, SignalsBaseApiTestCase):
 
     def test_get_pdf_signal_does_not_exists(self):
         self.client.force_authenticate(user=self.sia_read_write_user)
-        response = self.client.get(path=reverse(
-            'v1:signal-pdf-download',
-            kwargs={'pk': 999})
-        )
+        url = '/signals/v1/private/signals/{}/pdf'.format(999)
+        response = self.client.get(path=url)
 
         self.assertEqual(response.status_code, 404)
 
     def test_get_pdf_signal_not_loggedin(self):
         self.client.logout()
-
-        response = self.client.get(path=reverse(
-            'v1:signal-pdf-download',
-            kwargs={'pk': 999})
-        )
+        url = '/signals/v1/private/signals/{}/pdf'.format(999)
+        response = self.client.get(path=url)
 
         self.assertEqual(response.status_code, 401)
