+++ api/app/tox.ini
@@ -42,6 +42,7 @@ omit =
     */migrations/*
     signals/settings/*
     signals/wsgi.py
+    */urls.py
 
 [coverage:report]
 fail_under = 90
